move common_get_hf_file to common.cpp
